+++ x/slashing/genesis.go
@@ -11,7 +11,7 @@ import (
 func InitGenesis(ctx sdk.Context, keeper Keeper, stakingKeeper types.StakingKeeper, data types.GenesisState) {
 	stakingKeeper.IterateValidators(ctx,
 		func(index int64, validator exported.ValidatorI) bool {
-			keeper.addPubkey(ctx, validator.GetConsPubKey())
+			keeper.AddPubkey(ctx, validator.GetConsPubKey())
 			return false
 		},
 	)
@@ -30,20 +30,18 @@ func InitGenesis(ctx sdk.Context, keeper Keeper, stakingKeeper types.StakingKeep
 			panic(err)
 		}
 		for _, missed := range array {
-			keeper.setValidatorMissedBlockBitArray(ctx, address, missed.Index, missed.Missed)
+			keeper.SetValidatorMissedBlockBitArray(ctx, address, missed.Index, missed.Missed)
 		}
 	}
 
-	keeper.paramspace.SetParamSet(ctx, &data.Params)
+	keeper.SetParams(ctx, data.Params)
 }
 
 // ExportGenesis writes the current store values
 // to a genesis file, which can be imported again
 // with InitGenesis
 func ExportGenesis(ctx sdk.Context, keeper Keeper) (data types.GenesisState) {
-	var params types.Params
-	keeper.paramspace.GetParamSet(ctx, &params)
-
+	params := keeper.GetParams(ctx)
 	signingInfos := make(map[string]types.ValidatorSigningInfo)
 	missedBlocks := make(map[string][]types.MissedBlock)
 	keeper.IterateValidatorSigningInfos(ctx, func(address sdk.ConsAddress, info types.ValidatorSigningInfo) (stop bool) {
@@ -60,9 +58,5 @@ func ExportGenesis(ctx sdk.Context, keeper Keeper) (data types.GenesisState) {
 		return false
 	})
 
-	return types.GenesisState{
-		Params:       params,
-		SigningInfos: signingInfos,
-		MissedBlocks: missedBlocks,
-	}
+	return types.NewGenesisState(params, signingInfos, missedBlocks)
 }
