+++ x/slashing/keeper/params.go
@@ -53,3 +53,8 @@ func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
 	k.paramspace.GetParamSet(ctx, &params)
 	return params
 }
+
+// SetParams sets the slashing parameters to the param space.
+func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
+	k.paramspace.SetParamSet(ctx, &params)
+}
